move sitter profile into sep folder
