+++ src/components/index.js
@@ -21,12 +21,12 @@ export { default as SitterAccountSidebar } from './ui/SitterAccountSidebar.jsx';
 export { default as DiscoverSitters } from './sitters/DiscoverSitters.jsx';
 export { default as SitterProfile } from './profiles/sitters/SitterProfile.jsx';
 export { default as UserProfile } from './profiles/UserProfile.jsx';
-export { default as EditSitterProfile } from './users/EditSitterProfile.jsx';
+export { default as EditSitterProfile } from './users/sitters/EditSitterProfile.jsx';
 export { default as EditUserProfile } from './users/EditUserProfile.jsx';
 export { default as EditUserPets } from './users/EditUserPets.jsx';
-export { default as EditSitterPetPrefs } from './users/EditSitterPetPrefs.jsx';
-export { default as SitterOverview } from './users/SitterOverview.jsx';
-export { default as SitterBookings } from './users/SitterBookings.jsx';
+export { default as EditSitterPetPrefs } from './users/sitters/EditSitterPetPrefs.jsx';
+export { default as SitterOverview } from './users/sitters/SitterOverview.jsx';
+export { default as SitterBookings } from './users/sitters/SitterBookings.jsx';
 export { default as EditUserAccess } from './users/EditUserAccess.jsx';
 export { default as About } from './ui/About.jsx';
 export { default as UserSocialView } from './profiles/users/UserSocialView.jsx';
@@ -40,3 +40,4 @@ export { default as UserFriendsSidebar } from './profiles/users/UserFriendsSideb
 export { default as BookingRequestConfirmation } from './profiles/BookingRequestConfirmation.jsx';
 export { default as BookingSuccess } from './ui/BookingSuccess.jsx';
 export { default as Chat } from './ui/Chat.jsx';
+export { default as UserPetDetails } from './profiles/users/UserPetProfiles.jsx';
